qcombinator-backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Document Processing
pypdf==3.17.1
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.1.3
numpy==1.24.3

# AI/LLM Integration
openai==1.3.7
anthropic==0.7.7
langchain==0.0.340
tiktoken==0.5.1

# File Storage
boto3==1.29.7
minio==7.2.0

# Task Queue & Caching
celery==5.3.4
redis==5.0.1

# Data Validation
pydantic==2.5.2
email-validator==2.1.0

# PDF Generation
reportlab==4.0.7
weasyprint==60.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utils
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2

# Monitoring
prometheus-client==0.19.0